+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Update `sham-ui` dependencies
 * Refactoring store service in admin-panel [#84](https://github.com/sham-ui/sham-ui-cli/issues/84) 
 * Fix a tests
+* Make `settings` page in admin panel lazy
 
 ## 4.3.0
 * Add template for `styled-component` [#3](https://github.com/sham-ui/sham-ui-cli/issues/3)